Add `AbstractFormOptionAddForm`
