+++ wcfsetup/install/files/lib/acp/form/AbstractCustomOptionForm.class.php
@@ -15,7 +15,7 @@ use wcf\system\WCF;
  * @copyright   2001-2019 WoltLab GmbH
  * @license GNU Lesser General Public License <http://opensource.org/licenses/lgpl-license.php>
  * @since       3.1
- * @deprecated 6.2 Use `IFormOption` instead
+ * @deprecated 6.2 Use `AbstractFormOptionAddForm` instead
  */
 abstract class AbstractCustomOptionForm extends AbstractAcpForm
 {
