+++ wcfsetup/install/files/lib/acp/form/ContactOptionAddForm.class.php
@@ -5,21 +5,11 @@ namespace wcf\acp\form;
 use wcf\data\contact\option\ContactOption;
 use wcf\data\contact\option\ContactOptionAction;
 use wcf\data\contact\option\ContactOptionList;
-use wcf\data\IStorableObject;
 use wcf\form\AbstractFormBuilderForm;
-use wcf\system\form\builder\data\processor\CustomFormDataProcessor;
 use wcf\system\form\builder\field\BooleanFormField;
-use wcf\system\form\builder\field\dependency\ValueFormFieldDependency;
-use wcf\system\form\builder\field\IFormField;
 use wcf\system\form\builder\field\MultilineTextFormField;
-use wcf\system\form\builder\field\SelectFormField;
 use wcf\system\form\builder\field\ShowOrderFormField;
 use wcf\system\form\builder\field\TextFormField;
-use wcf\system\form\builder\IFormDocument;
-use wcf\system\form\option\FormOptionHandler;
-use wcf\system\form\option\SharedConfigurationFormFields;
-use wcf\system\WCF;
-use wcf\util\JSON;
 
 /**
  * Shows the contact option add form.
@@ -31,7 +21,7 @@ use wcf\util\JSON;
  *
  * @extends AbstractFormBuilderForm<ContactOption>
  */
-class ContactOptionAddForm extends AbstractFormBuilderForm
+class ContactOptionAddForm extends AbstractFormOptionAddForm
 {
     /**
      * @inheritDoc
@@ -73,54 +63,11 @@ class ContactOptionAddForm extends AbstractFormBuilderForm
                 ->languageItemPattern('wcf.contact.optionDescription\d+'),
             ShowOrderFormField::create('showOrder')
                 ->options($this->getContactOptions()),
-            SelectFormField::create('optionType')
-                ->label('wcf.acp.customOption.optionType')
-                ->immutable($this->formAction != 'create')
-                ->options($this->getAvailableOptionTypes())
-                ->required(),
+            $this->getOptionTypeFormField(),
             ...$this->getSharedConfigurationFormFields(),
-            BooleanFormField::create('required')
-                ->label('wcf.acp.customOption.required'),
             BooleanFormField::create('isDisabled')
                 ->label('wcf.acp.customOption.isDisabled'),
         ]);
-    }
-
-    #[\Override]
-    public function finalizeForm()
-    {
-        parent::finalizeForm();
-
-        $this->form->getDataHandler()->addProcessor(
-            new CustomFormDataProcessor(
-                'saveOptionProcessor',
-                function (IFormDocument $document, array $parameters) {
-                    $configurationData = [];
-
-                    foreach ($this->getConfigurationFormFieldIds() as $parameter) {
-                        if (isset($parameters['data'][$parameter])) {
-                            $configurationData[$parameter] = $parameters['data'][$parameter];
-                            unset($parameters['data'][$parameter]);
-                        }
-                    }
-
-                    if ($configurationData !== []) {
-                        $parameters['data']['configurationData'] = JSON::encode($configurationData);
-                    }
-
-                    return $parameters;
-                },
-                function (IFormDocument $document, array $data, IStorableObject $object) {
-                    \assert($object instanceof ContactOption);
-
-                    if ($object->configurationData) {
-                        $data = \array_merge($data, JSON::decode($object->configurationData));
-                    }
-
-                    return $data;
-                }
-            )
-        );
     }
 
     /**
@@ -133,70 +80,5 @@ class ContactOptionAddForm extends AbstractFormBuilderForm
         $optionList->readObjects();
 
         return \array_map(static fn($option) => $option->getTitle(), $optionList->getObjects());
-    }
-
-    /**
-     * @return array<string, string>
-     */
-    private function getAvailableOptionTypes(): array
-    {
-        $optionTypes = \array_map(fn($option) => $option->getTitle(), FormOptionHandler::getInstance()->getOptions());
-
-        $collator = new \Collator(WCF::getLanguage()->getLocale());
-        \uasort(
-            $optionTypes,
-            static fn(string $a, string $b) => $collator->compare($a, $b)
-        );
-
-        return $optionTypes;
-    }
-
-    /**
-     * @return IFormField[]
-     */
-    private function getSharedConfigurationFormFields(): array
-    {
-        $sharedConfigurationFormFields = new SharedConfigurationFormFields();
-        $matrix = [];
-
-        foreach (FormOptionHandler::getInstance()->getOptions() as $option) {
-            foreach ($option->getConfigurationFormFields() as $formFieldId) {
-                if (!isset($matrix[$formFieldId])) {
-                    $matrix[$formFieldId] = [];
-                }
-
-                $matrix[$formFieldId][] = $option->getId();
-            }
-        }
-
-        $formFields = [];
-
-        foreach ($matrix as $formFieldId => $dependencies) {
-            $formField = $sharedConfigurationFormFields->getFormField($formFieldId);
-            $formField->addDependency(
-                ValueFormFieldDependency::create($formFieldId . 'OptionTypeDependency')
-                    ->fieldId('optionType')
-                    ->values($dependencies)
-            );
-            $formFields[] = $formField;
-        }
-
-        return $formFields;
-    }
-
-    /**
-     * @return string[]
-     */
-    private function getConfigurationFormFieldIds(): array
-    {
-        $ids = [];
-
-        foreach (FormOptionHandler::getInstance()->getOptions() as $option) {
-            foreach ($option->getConfigurationFormFields() as $formFieldId) {
-                $ids[] = $formFieldId;
-            }
-        }
-
-        return \array_unique($ids);
     }
 }
