+++ wcfsetup/install/files/lib/system/form/option/AbstractFormOption.class.php
@@ -20,7 +20,7 @@ abstract class AbstractFormOption implements IFormOption
     #[\Override]
     public function getConfigurationFormFields(): array
     {
-        return [];
+        return ['required'];
     }
 
     #[\Override]
