+++ wcfsetup/install/files/lib/system/form/option/CheckboxesFormOption.class.php
@@ -53,9 +53,7 @@ class CheckboxesFormOption extends AbstractFormOption
     #[\Override]
     public function getConfigurationFormFields(): array
     {
-        return [
-            'selectOptions'
-        ];
+        return ['selectOptions', 'required'];
     }
 
     #[\Override]
