+++ wcfsetup/install/lang/en.xml
@@ -4081,6 +4081,8 @@ Attachments:
 		<item name="wcf.form.option.shared.minValue"><![CDATA[Minimum]]></item>
 		<item name="wcf.form.option.shared.maxValue"><![CDATA[Maximum]]></item>
 		<item name="wcf.form.option.shared.selectOptions"><![CDATA[Options]]></item>
+		<item name="wcf.form.option.shared.required"><![CDATA[This is a required field]]></item>
+		<item name="wcf.form.option.optionType"><![CDATA[Option Type]]></item>
 	</category>
 	<category name="wcf.image">
 		<item name="wcf.image.coverPhoto"><![CDATA[Cover Photo]]></item>
